feat: change project structures
