+++ src/commands/params.ts
@@ -1,6 +1,28 @@
 import { CommandInteraction, EmbedBuilder } from 'discord.js';
 import Command from './commands';
-import { getParamsText } from '../utils/params';
+import { customErrorHandler } from '../common/error';
+import envs from '../common/envs';
+import { getRequest } from '../common/utils';
+
+const { ENDPOINT } = envs;
+
+const getParamsText = async (taskId: string): Promise<string | null> => {
+  try {
+    const { isSuccess, data } = await getRequest(`${ENDPOINT}/tasks/${taskId}/params`);
+    if (!isSuccess) {
+      return null;
+    }
+    let paramsText = '';
+    Object.entries(data).forEach(([paramName, paramValue]) => {
+      paramsText += `> **${paramName}**\n> ${paramValue}\n`;
+    });
+
+    return paramsText;
+  } catch (error) {
+    const errorMessage = customErrorHandler(error);
+    return errorMessage;
+  }
+};
 
 const params = async (interaction: CommandInteraction) => {
   if (!interaction || interaction.user.bot || !interaction.isChatInputCommand() || !interaction.guildId) return;
